requirements.txt
requests 
mysql-connector-python 
sqlalchemy 
flask 
flask_sqlalchemy
Flask-JWT-Extended
python-dotenv